[hive] Improve hive database is exists yet create/drop database

+++ paimon-hive/paimon-hive-catalog/src/main/java/org/apache/paimon/hive/HiveCatalog.java
@@ -208,10 +208,7 @@ public class HiveCatalog extends AbstractCatalog {
     @Override
     protected boolean databaseExistsImpl(String databaseName) {
         try {
-            client.getDatabase(databaseName);
-            return true;
-        } catch (NoSuchObjectException e) {
-            return false;
+            return getDatabase(databaseName) != null;
         } catch (TException e) {
             throw new RuntimeException(
                     "Failed to determine if database " + databaseName + " exists", e);
@@ -221,13 +218,16 @@ public class HiveCatalog extends AbstractCatalog {
     @Override
     protected void createDatabaseImpl(String name) {
         try {
-            Path databasePath = newDatabasePath(name);
-            locationHelper.createPathIfRequired(databasePath, fileIO);
+            Database database = getDatabase(name);
+            if (database == null) {
+                Path databasePath = newDatabasePath(name);
+                locationHelper.createPathIfRequired(databasePath, fileIO);
 
-            Database database = new Database();
-            database.setName(name);
-            locationHelper.specifyDatabaseLocation(databasePath, database);
-            client.createDatabase(database);
+                database = new Database();
+                database.setName(name);
+                locationHelper.specifyDatabaseLocation(databasePath, database);
+                client.createDatabase(database);
+            }
         } catch (TException | IOException e) {
             throw new RuntimeException("Failed to create database " + name, e);
         }
@@ -236,12 +236,25 @@ public class HiveCatalog extends AbstractCatalog {
     @Override
     protected void dropDatabaseImpl(String name) {
         try {
-            Database database = client.getDatabase(name);
-            String location = locationHelper.getDatabaseLocation(database);
-            locationHelper.dropPathIfRequired(new Path(location), fileIO);
-            client.dropDatabase(name, true, false, true);
+            Database database = getDatabase(name);
+            if (database != null) {
+                String location = locationHelper.getDatabaseLocation(database);
+                locationHelper.dropPathIfRequired(new Path(location), fileIO);
+                client.dropDatabase(name, true, false, true);
+            }
         } catch (TException | IOException e) {
             throw new RuntimeException("Failed to drop database " + name, e);
+        }
+    }
+
+    private Database getDatabase(String databaseName) throws TException {
+        try {
+            return client.getDatabase(databaseName);
+        } catch (NoSuchObjectException e) {
+            return null;
+        } catch (TException e) {
+            throw new RuntimeException(
+                    "Failed to determine if database " + databaseName + " exists", e);
         }
     }
